backend/requirements.txt
Flask
Flask-PyMongo
bcrypt
PyJWT
flasgger
flask-cors
pymongo
